requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
tqdm>=4.64.0

# LLM and AI
openai>=1.0.0
chromadb>=0.4.0

# Database
neo4j>=5.0.0

# Web UI
streamlit>=1.28.0

# Data processing and analysis
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0

# Utilities
python-dotenv>=0.19.0
openpyxl>=3.0.0
pathlib2>=2.3.0

# Optional: For Ollama integration
# ollama>=0.1.0 